tests: moved parts of random nodes test to external package
(cherry picked from commit a32a029f01863b31948330641363e623e3ccf71e)
